mini-services/lumibot-service/requirements.txt
lumibot>=4.4.53
ccxt>=4.5.42
pandas>=3.0.1
numpy>=2.4.2
scipy>=1.17.1
yfinance>=1.2.0
flask>=3.1.3
flask-cors>=4.0.0
python-dotenv>=1.2.2
apscheduler>=3.11.2
